requirements.txt
alembic==1.14.0
aniso8601==9.0.1
bcrypt==4.2.1
blinker==1.9.0
click==8.1.7
exceptiongroup==1.2.2
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Flask-Session==0.5.0
greenlet==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.8
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
PyJWT==2.10.1
pytest==8.3.4
pytest-cov==6.0.0
python-dotenv==1.0.1
pytz==2024.2
redis==5.0.1
six==1.17.0
SQLAlchemy==2.0.36
tomli==2.2.1
typing_extensions==4.12.2
Werkzeug==3.1.3
WTForms==3.2.1
